\name{CommonForecastingErrors}
\alias{CommonForecastingErrors}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Common Forecasting Errors
}
\description{
Common Forecasting Errors
}
\usage{
CommonForecastingErrors(TestdataY,ForecastingF,

epsilon=10^-4,na.rm=TRUE,stepsize=1,digits)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{TestdataY}{
[1:n] vector of test data
}
  \item{ForecastingF}{
[1:n] vector of forecast
}
  \item{epsilon}{
Optionalm epsilon defining when  zero values should be approximated
}
  \item{na.rm}{
Optional, removing missing values
}
  \item{stepsize}{
Optional, for mase
}
  \item{digits}{
Optional, number of digits for the output values to which the results will be scientifically rounded to (and not by R logic)
}
}
\details{
to be filled
}
\value{
todo
}
\references{
to be filled
}
\author{
Michael Thrun
}
\note{
to be filled
}


\examples{
#ToDo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{CommonForecastingErrors}% use one of  RShowDoc("KEYWORDS")

