\name{EchoStateANN_forecasting}
\alias{EchoStateANN_forecasting}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
EchoStateANN_forecasting
}
\description{
EchoStateANN_forecasting
}
\usage{
EchoStateANN_forecasting(Datavector, Percentage = 90, 

Forecasthorizon = 1, Seasonality = 28, Scaled = TRUE,

Time, PlotIt = FALSE, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Datavector}{
Regular (equidistant) time series data as numerical vector.
}
  \item{Percentage}{
percentage of the split between training and test data
}
  \item{Forecasthorizon}{
Number of time steps to forecast into the future
}
  \item{Seasonality}{
main saisonality of data, is used for generating batches of data
}
  \item{Scaled}{
TRUE: automatic scaling
}
  \item{Time}{
Optional, Time vector in the length of data
}
  \item{PlotIt}{
Optional, A logical determining whether or not to plot the forecast in comparison to the validation set.
}
  \item{\dots}{
further arguments for \code{\link[echos]{train_esn}}
}
}
\details{
requires interally also \pkg{tsibble} 
}
\value{
List of 
\item{Model}{ANN model generated by \pkg{echos}}
\item{FitStats}{Output of \code{forecast_esn} in [echos]}
\item{Forecast}{Forecast generated by the ANN model where we put in the last portion of the training set of length \code{forecast_length} as data to predict from. The test data stays untouched.}
\item{TestData}{[(k+1):n], the part of Response not used in the model}
\item{TestTime}{[(k+1):n], time of response not used in the model}
\item{TrainData}{[1:k], the part of Response used in the model}
\item{TrainTime}{[1:k], time of Training data if geven}
\item{ForecastTrain}{[1:k], forecasted value using TrainData}
}

\references{
Jaeger H. (2003): Adaptive nonlinear system identification with echo state networks. In Advances in Neural Information Processing Systems 15, S. Becker, S. Thrun, K. Obermayer (Eds), (MIT Press, Cambridge, MA, 2003) pp. 593-600 
}
\author{
Michael Thrun
}
\note{
devtools::install_github("ahaeusser/echos",dependencies = T)
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link[echos]{train_esn}}
}

\examples{
\donttest{
if(requireNamespace('ggfortify')){
library(ggfortify)
x=fortify(datasets::sunspot.month)
#Example for a bad forecast
results=EchoStateANN_forecasting(Datavector = x$Data,Time = x$Index,Seasonality=12)
}
}
}

\keyword{ANN}
\keyword{echostate}

