\name{GermanHolidays}
\alias{GermanHolidays}
\docType{data}
\title{
GermanHolidays
}
\description{
GermanHolidays
}
\usage{data("GermanHolidays")}

\details{
German Holidays from "1988-01-01 UTC" until "2038-12-26 UTC" as a dataframe consisting of a \code{POSIXct} Time and a \code{character} description
}

\examples{
data(GermanHolidays)
str(GermanHolidays)
}
\keyword{datasets}
