\name{GermanPrimes}
\alias{GermanPrimes}
\docType{data}
\title{
GermanPrimes
}
\description{
GermanPrimes
}
\usage{data("GermanPrimes")}
\format{
  A data frame with 324 observations on the following 8 variables.
  \describe{
    \item{\code{ISIN}}{a character vector}
    \item{\code{Trading.Symbol}}{a character vector}
    \item{\code{Company}}{a character vector}
    \item{\code{Sector}}{a character vector}
    \item{\code{Subsector}}{a character vector}
    \item{\code{Country}}{a character vector}
    \item{\code{Instrument.Exchange}}{a character vector}
    \item{\code{Yahoo}}{a character vector}
  }
}
\details{
to be filled
}
\source{
to be filled
}
\references{
to be filled
}
\examples{
data(GermanPrimes)
## maybe str(GermanPrimes) ; plot(GermanPrimes) ...
}
\keyword{datasets}
