\name{RelativeDifferenceTS}
\alias{RelativeDifferenceTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
RelativeDifferenceTS
}
\description{
RelativeDifference  for Time series
}
\usage{
RelativeDifferenceTS(TS1, TS2, na.rm = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{TS1}{
[1:n] time series data vector
}
  \item{TS2}{
[1:n] time series data vector
}
  \item{na.rm}{
TRUE: interpolate NaN
}
}

\value{
Relative difference [1:n] between two TS
}

\author{
Michael Thrun
}

\keyword{RelativeDifferenceTS}% use one of  RShowDoc("KEYWORDS"
