\name{Sales}
\alias{Sales}
\docType{data}
\title{
Sales data
}
\description{
Randomly generated Sales data
}
\usage{data("Sales")}
\format{
  The format is:
 num [1:72] 150 105 145 155 200 240 265 235 185 195 ...
}
\details{
Randomly generated Sales data
}


\examples{
data(Sales)
## maybe str(Sales) ; plot(Sales) ...
}
\keyword{datasets}
