\name{TempMelbourneAustralia}
\alias{TempMelbourneAustralia}
\docType{data}
\title{
Temperature in Melbourne
}
\description{
Daily temperatures over 10 years (1981-1990) in the city Melbourne, Australia.
}
\usage{data("TempMelbourneAustralia")}
\format{
  A data frame with 3650 observations on the following 2 variables.
  \describe{
    \item{\code{Time}}{a Date}
    \item{\code{Temp}}{a numeric vector}
  }
}
\details{
The units are in degrees Celsius and there are 3650 observations. Given is the daily minimum temperature.
}
\source{
The source of the data is credited as the Australian Bureau of Meteorology.}
\references{
machinelearningmastery.com
}
\examples{
data(TempMelbourneAustralia)
}
\keyword{datasets}
