\name{Time2Classification}
\alias{Time2Classification}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Time to Classification
}
\description{
Generates Classification out of non unqiue time stamps
}
\usage{
Time2Classification(OrderedTimeChar)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{OrderedTimeChar}{
[1:n] vector
}
}
\details{
Transfers \code{OrderedTimeChar} to a character vector
}
\value{
[1:n] Classification of k time stamps
}

\author{
Michael Thrun
}

\examples{
##todo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Time2Classification}% use one of  RShowDoc("KEYWORDS")
