\name{WeeklySalesUniformRandom}
\alias{WeeklySalesUniformRandom}
\docType{data}
\title{
WeeklySalesUniformRandom
}
\description{
WeeklySalesUniformRandom
}
\usage{data("WeeklySalesUniformRandom")}

\details{
WeeklySalesUniformRandom
}
\source{
artificially generated

}

\examples{
data(WeeklySalesUniformRandom)
str(WeeklySalesUniformRandom)
}
\keyword{datasets}
