\name{WriteDates}
\alias{WriteDates}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
WriteDates
}
\description{
WriteDates
}
\usage{
WriteDates(FileName, TSdata, Key = c(), OutDirectory = getwd(), Comments = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{FileName}{
under development
}
  \item{TSdata}{
under development
}
  \item{Key}{
under development
}
  \item{OutDirectory}{
under development
}
  \item{Comments}{
under development
}
}
\details{
under development
}
\value{
under development
}

\author{
Michael Thrun
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{ReadDates}}
}
\examples{
#todo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ReadDates}% use one of  RShowDoc("KEYWORDS")
