\name{ReadDates}
\alias{ReadDates}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ReadDates
}
\description{
under development
}
\usage{
ReadDates(FileName = NULL, InDirectory = getwd(), SilentComments = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{FileName}{
under development
}
  \item{InDirectory}{
under development
}
  \item{SilentComments}{
under development
}
}
\details{
under development
}
\value{
data frame
}
\references{
under development
}
\author{
Michael Thrun
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{WriteDates}}
}
\examples{
##todo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{WriteDates}% use one of  RShowDoc("KEYWORDS")